test: Wait for flush to happen in IntegrationCodecClient::flushWrite
Unfortunately implementing this by tracking output buffer watermarks requires changes to the connection interface so that the read and write buffer watermarks can be controlled independently. This will be specially true after #14054 which adjusts read size based on the read buffer high-watermark. This change also depends on #14333 since without it is not possible to get a true 1 byte output buffer highwatermark.

Signed-off-by: Antonio Vicente <avd@google.com>